fix: split test into 2 files
Summary of changes:
- Split seldon server tests into separate file.
